+++ ui/packages/app/web/src/components/ui/Navbar.tsx
@@ -12,7 +12,7 @@
 // limitations under the License.
 
 import {Disclosure} from '@headlessui/react';
-import {MenuIcon, XIcon} from '@heroicons/react/outline';
+import {Bars3Icon, XMarkIcon} from '@heroicons/react/20/solid';
 import {Parca, ParcaSmall} from '@parca/icons';
 import GitHubButton from 'react-github-btn';
 import cx from 'classnames';
@@ -69,9 +69,9 @@ const Navbar = () => {
                 <Disclosure.Button className="inline-flex items-center justify-center p-2 rounded-md text-gray-400 hover:text-white hover:bg-gray-700 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white">
                   <span className="sr-only">Open main menu</span>
                   {open ? (
-                    <XIcon className="block h-6 w-6" aria-hidden="true" />
+                    <XMarkIcon className="block h-6 w-6" aria-hidden="true" />
                   ) : (
-                    <MenuIcon className="block h-6 w-6" aria-hidden="true" />
+                    <Bars3Icon className="block h-6 w-6" aria-hidden="true" />
                   )}
                 </Disclosure.Button>
               </div>