+++ ui/packages/shared/profile/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.112](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.111...@parca/profile@0.16.112) (2023-02-09)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.111](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.109...@parca/profile@0.16.111) (2023-02-09)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.110](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.109...@parca/profile@0.16.110) (2023-02-08)
+
+**Note:** Version bump only for package @parca/profile
+
+## 0.16.109 (2023-02-08)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.108](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.107...@parca/profile@0.16.108) (2023-02-08)
+
+**Note:** Version bump only for package @parca/profile
+
+## 0.16.107 (2023-02-07)
+
+**Note:** Version bump only for package @parca/profile
+
 ## [0.16.106](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.101...@parca/profile@0.16.106) (2023-02-06)
 
 **Note:** Version bump only for package @parca/profile